Started working on vertex painting example
Example script will utilize parts from utils and blender.utils to color
vertices in Blender. Current blender README has an example script
using an operator, but it should be possible to set the color to the
loop indices as well. Last time I tried this it crashed Blender, but will
try it again in this script.